# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CAB Booking System API",
    "version": "1.0.0",
    "description": "Aggregated OpenAPI snapshot for the current CAB Booking System repo state."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "API Gateway"
    }
  ],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register a new account",
        "requestBody": { "$ref": "#/components/requestBodies/RegisterRequest" },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Validation error" } }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login",
        "requestBody": { "$ref": "#/components/requestBodies/LoginRequest" },
        "responses": { "200": { "description": "Authenticated" }, "401": { "description": "Invalid credentials" } }
      }
    },
...
    "/api/users/{userId}": {
      "get": {
        "tags": ["users"],
        "summary": "Get user by id",
        "parameters": [{ "$ref": "#/components/parameters/UserIdParam" }],
        "responses": { "200": { "description": "User" }, "403": { "description": "Forbidden" } }
      }
    },
...
    "/api/bookings": {
      "post": {
        "tags": ["bookings"],
        "summary": "Create booking",
        "requestBody": { "$ref": "#/components/requestBodies/BookingCreateRequest" },
        "responses": { "201": { "description": "Created" }, "200": { "description": "Replay or handled scenario" }, "400": { "description": "Validation error" }, "422": { "description": "Schema validation error" } }
      }
    },
    "/api/bookings/{id}": {
      "get": {
        "tags": ["bookings"],
        "summary": "Get booking by id",
        "parameters": [{ "$ref": "#/components/parameters/BookingIdParam" }],
        "responses": { "200": { "description": "Booking" } }
      }
    },
...
    "/api/rides/{rideId}": {
      "get": {
        "tags": ["rides"],
        "summary": "Get ride details",
        "parameters": [{ "$ref": "#/components/parameters/RideIdParam" }],
        "responses": { "200": { "description": "Ride" } }
      }
    },
...
    "/api/payments": {
      "post": {
        "tags": ["payments"],
        "summary": "Create payment",
        "requestBody": { "$ref": "#/components/requestBodies/PaymentCreateRequest" },
        "responses": { "201": { "description": "Created" }, "200": { "description": "Idempotent replay" } }
      }
    },
...
    "/api/payments/{paymentId}": {
      "get": {
        "tags": ["payments"],
        "summary": "Get payment by id",
        "parameters": [{ "$ref": "#/components/parameters/PaymentIdParam" }],
        "responses": { "200": { "description": "Payment" } }
      }
    },
...
  },
  "components": {
    "parameters": {
      "UserIdParam": {
        "name": "userId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
      "RideIdParam": {
        "name": "rideId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
      "BookingIdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
      "PaymentIdParam": {
        "name": "paymentId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      }
    },
    "requestBodies": {
      "RegisterRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["email", "phone", "password"],
              "properties": {
                "email": { "type": "string" },
                "phone": { "type": "string" },
                "password": { "type": "string" },
                "firstName": { "type": "string" },
                "lastName": { "type": "string" },
                "role": { "type": "string", "enum": ["customer", "driver", "admin"] }
              }
            }
          }
        }
      },
      "LoginRequest": {
        "required": true,
        "content": { "application/json": { "schema": { "type": "object", "required": ["email", "password"], "properties": { "email": { "type": "string" }, "password": { "type": "string" } } } } }
      },
      "BookingCreateRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["customerId", "pickupLocation", "dropoffLocation"],
              "properties": {
                "customerId": { "type": "string" },
                "pickupLocation": { "$ref": "#/components/schemas/Location" },
                "dropoffLocation": { "$ref": "#/components/schemas/Location" },
                "paymentMethod": { "type": "string", "enum": ["CASH", "CARD", "WALLET"] },
                "autoAssign": { "type": "boolean" },
                "searchRadiusKm": { "type": "number" },
                "idempotencyKey": { "type": "string" }
              }
            }
          }
        }
      },
...
      "PaymentCreateRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["rideId", "amount", "userId"],
              "properties": {
                "rideId": { "type": "string" },
                "amount": { "type": "number" },
                "userId": { "type": "string" },
                "driverId": { "type": "string" },
                "method": { "type": "string" },
                "provider": { "type": "string" },
                "idempotencyKey": { "type": "string" }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "Location": {
        "type": "object",
        "required": ["address", "latitude", "longitude"],
        "properties": {
          "address": { "type": "string" },
          "latitude": { "type": "number" },
          "longitude": { "type": "number" }
        }
      }
    }
  }
}